+++ packages/core/src/index.ts
@@ -8,4 +8,3 @@ export * from './plugins/use-error-handler';
 export * from './plugins/use-extend-context';
 export * from './plugins/use-payload-formatter';
 export * from './plugins/use-masked-errors';
-export * from './plugins/use-context-per-subscription-event';
